requirement.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
google-generativeai
numpy
pandas
scikit-learn
